api complete

+++ routes/api.php
@@ -1,16 +1,17 @@
 <?php
 
-use App\Http\Controllers\ApiController;
+use App\Http\Controllers\AdminApiController;
 use App\Http\Controllers\StudentApiController;
+use App\Http\Controllers\TeacherApiController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 
 //for admin
-Route::post('/register', [ApiController::class, 'register']);
-Route::post('/login', [ApiController::class, 'login']);
-Route::group(['prefix' => 'student', 'middleware' => 'auth:api'], function () {
-    Route::post('logout', [ApiController::class, 'logout']);
+Route::post('/admin/login', [AdminApiController::class, 'login']);
+Route::group(['prefix' => 'admin', 'middleware' => 'auth:api'], function () {
+    Route::post('logout', [AdminApiController::class, 'logout']);
+    Route::get('profile', [AdminApiController::class, 'profile']);
 });
 
 // for student
@@ -19,6 +20,16 @@ Route::post('/student/login', [StudentApiController::class, 'login']);
 Route::group(['prefix' => 'student', 'middleware' => 'auth:api'], function () {
     Route::post('logout', [StudentApiController::class, 'logout']);
     Route::get('profile', [StudentApiController::class, 'profile']);
-    Route::put('update/{userId}', [StudentApiController::class, 'updateStudent']);
-    Route::delete('delete', [StudentApiController::class, 'deleteStudent']);
+    Route::patch('update/{userId}', [StudentApiController::class, 'updateStudent']);
+    Route::delete('delete/{userId}', [StudentApiController::class, 'deleteStudent']);
+});
+
+// for Teacher
+Route::post('/teacher/register', [TeacherApiController::class, 'register']);
+Route::post('/teacher/login', [TeacherApiController::class, 'login']);
+Route::group(['prefix' => 'teacher', 'middleware' => 'auth:api'], function () {
+    Route::post('logout', [TeacherApiController::class, 'logout']);
+    Route::get('profile', [TeacherApiController::class, 'profile']);
+    Route::patch('update/{userId}', [TeacherApiController::class, 'update']);
+    Route::delete('delete/{userId}', [TeacherApiController::class, 'delete']);
 });
